requirements_2019_and_newer_model.txt
biopython>=1.69
tensorflow==1.12.0
numpy==1.16.0
h5py==2.8.0
Keras==2.2.4
scikit-learn>=0.20.10
